# Seckill-Core/api/Seckill-Core/v1/Seckill-Core_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.7.0
// - protoc             v4.25.0--rc1
// source: Seckill-Core/v1/Seckill-Core.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationSeckill-CoreCreateUser = "/api.shortUrlX.v1.Seckill-Core/CreateUser"
const OperationSeckill-CoreGetUser = "/api.shortUrlX.v1.Seckill-Core/GetUser"
const OperationSeckill-CoreGetUserByName = "/api.shortUrlX.v1.Seckill-Core/GetUserByName"

type Seckill-CoreHTTPServer interface {
	CreateUser(context.Context, *CreateUserRequest) (*CreateUserReply, error)
	GetUser(context.Context, *GetUserRequest) (*GetUserReply, error)
	GetUserByName(context.Context, *GetUserByNameRequest) (*GetUserByNameReply, error)
}

func RegisterSeckill-CoreHTTPServer(s *http.Server, srv Seckill-CoreHTTPServer) {
	r := s.Route("/")
	r.POST("/create_user", _Seckill-Core_CreateUser0_HTTP_Handler(srv))
	r.GET("/get_user_info", _Seckill-Core_GetUser0_HTTP_Handler(srv))
	r.GET("/get_user_info_byname", _Seckill-Core_GetUserByName0_HTTP_Handler(srv))
}

func _Seckill-Core_CreateUser0_HTTP_Handler(srv Seckill-CoreHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateUserRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSeckill-CoreCreateUser)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateUser(ctx, req.(*CreateUserRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateUserReply)
		return ctx.Result(200, reply)
	}
}

func _Seckill-Core_GetUser0_HTTP_Handler(srv Seckill-CoreHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetUserRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSeckill-CoreGetUser)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetUser(ctx, req.(*GetUserRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetUserReply)
		return ctx.Result(200, reply)
	}
}

func _Seckill-Core_GetUserByName0_HTTP_Handler(srv Seckill-CoreHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetUserByNameRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSeckill-CoreGetUserByName)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetUserByName(ctx, req.(*GetUserByNameRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetUserByNameReply)
		return ctx.Result(200, reply)
	}
}

type Seckill-CoreHTTPClient interface {
	CreateUser(ctx context.Context, req *CreateUserRequest, opts ...http.CallOption) (rsp *CreateUserReply, err error)
	GetUser(ctx context.Context, req *GetUserRequest, opts ...http.CallOption) (rsp *GetUserReply, err error)
	GetUserByName(ctx context.Context, req *GetUserByNameRequest, opts ...http.CallOption) (rsp *GetUserByNameReply, err error)
}

type Seckill-CoreHTTPClientImpl struct {
	cc *http.Client
}

func NewSeckill-CoreHTTPClient(client *http.Client) Seckill-CoreHTTPClient {
	return &Seckill-CoreHTTPClientImpl{client}
}

func (c *Seckill-CoreHTTPClientImpl) CreateUser(ctx context.Context, in *CreateUserRequest, opts ...http.CallOption) (*CreateUserReply, error) {
	var out CreateUserReply
	pattern := "/create_user"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSeckill-CoreCreateUser))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *Seckill-CoreHTTPClientImpl) GetUser(ctx context.Context, in *GetUserRequest, opts ...http.CallOption) (*GetUserReply, error) {
	var out GetUserReply
	pattern := "/get_user_info"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSeckill-CoreGetUser))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *Seckill-CoreHTTPClientImpl) GetUserByName(ctx context.Context, in *GetUserByNameRequest, opts ...http.CallOption) (*GetUserByNameReply, error) {
	var out GetUserByNameReply
	pattern := "/get_user_info_byname"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSeckill-CoreGetUserByName))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}